cheri: Apply the tasklet.cocci script
Run the tasklet.cocci script to fix tasklet calling API.

Coccinelle can do cross-file analysis provided that all the
affected files are in a single invocation. Thus to catch all
matches of the script use the following hack:
- Determine all directories that contain potential matches,
  e.g. by grepping for file that contain the string "tasklet".
- For each directory <DIR> run
	spatch -sp-file scripts/coccinelle/cheri/tasklet.cocci <DIR>/*.[ch]

+++ drivers/atm/fore200e.c
@@ -1175,7 +1175,7 @@ fore200e_interrupt(int irq, void* dev)
 
 #ifdef FORE200E_USE_TASKLET
 static void
-fore200e_tx_tasklet(unsigned long data)
+fore200e_tx_tasklet(uintptr_t data)
 {
     struct fore200e* fore200e = (struct fore200e*) data;
     unsigned long flags;
@@ -1189,7 +1189,7 @@ fore200e_tx_tasklet(unsigned long data)
 
 
 static void
-fore200e_rx_tasklet(unsigned long data)
+fore200e_rx_tasklet(uintptr_t data)
 {
     struct fore200e* fore200e = (struct fore200e*) data;
     unsigned long    flags;
@@ -1938,8 +1938,10 @@ static int fore200e_irq_request(struct fore200e *fore200e)
 	   fore200e_irq_itoa(fore200e->irq), fore200e->name);
 
 #ifdef FORE200E_USE_TASKLET
-    tasklet_init(&fore200e->tx_tasklet, fore200e_tx_tasklet, (unsigned long)fore200e);
-    tasklet_init(&fore200e->rx_tasklet, fore200e_rx_tasklet, (unsigned long)fore200e);
+    tasklet_init(&fore200e->tx_tasklet, fore200e_tx_tasklet,
+                 (uintptr_t) fore200e);
+    tasklet_init(&fore200e->rx_tasklet, fore200e_rx_tasklet,
+                 (uintptr_t) fore200e);
 #endif
 
     fore200e->state = FORE200E_STATE_IRQ;
